# ai-control-core/acc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FluxForge ACC — AI Control Core",
    "description": "API for reading project files, searching code, and applying patches to FluxForge Studio. Used by ChatGPT Custom GPT as a failover when Claude Code is unavailable.",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://YOUR_NGROK_URL.ngrok-free.app",
      "description": "ACC server via ngrok tunnel"
    }
  ],
  "paths": {
...
    "/patch/apply": {
      "post": {
        "operationId": "applyPatch",
        "summary": "Apply a unified diff patch to the codebase",
        "description": "Submit a unified diff patch. ACC will: create branch, apply patch, run gates (locked paths + flutter analyze), and auto-merge into main if all gates pass.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["provider", "task_id", "patch"],
                "properties": {
                  "provider": {"type": "string", "enum": ["claude", "openai"], "description": "Which AI provider is submitting"},
                  "task_id": {"type": "string", "description": "Task ID (e.g. 'TASK_042')"},
                  "reason": {"type": "string", "description": "Short description of the change"},
                  "patch": {"type": "string", "description": "Unified diff format patch (must start with 'diff --git')"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patch result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {"type": "boolean"},
                    "merged": {"type": "boolean"},
                    "branch": {"type": "string"},
                    "changed_files": {"type": "array", "items": {"type": "string"}},
                    "blocked_files": {"type": "array", "items": {"type": "string"}},
                    "gate_results": {"type": "object"},
                    "error": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/task/create": {
      "post": {
        "operationId": "createTask",
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["task_id", "description"],
                "properties": {
                  "task_id": {"type": "string"},
                  "description": {"type": "string"},
                  "provider": {"type": "string", "default": "openai"},
                  "acceptance_criteria": {"type": "array", "items": {"type": "string"}}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {"type": "boolean"},
                    "task_id": {"type": "string"},
                    "error": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/task/close": {
      "post": {
        "operationId": "closeTask",
        "summary": "Close a task with result",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["task_id", "result"],
                "properties": {
                  "task_id": {"type": "string"},
                  "result": {"type": "string", "enum": ["PASS", "FAIL", "SKIPPED"]},
                  "notes": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task closed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {"type": "boolean"},
                    "task_id": {"type": "string"},
                    "error": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-acc-key"
      }
    }
  },
  "security": [
    {"apiKey": []}
  ]
}